+++ README.md
@@ -1,26 +1,24 @@
-# go-client-mongodb-realm
+# go-client-mongodb-atlas-app-services (WIP)
 [![PkgGoDev](https://pkg.go.dev/badge/go.mongodb.org/realm)](https://pkg.go.dev/go.mongodb.org/realm)
 
-A Go HTTP client for the [MongoDB Realm API](https://docs.mongodb.com/realm/admin/api/v3/).
-
-Note that Realm only supports the two most recent major versions of Go.
+A Go HTTP client for the [MongoDB Atlas App Services Admin API](https://www.mongodb.com/docs/atlas/app-services/admin/api/v3/).
 
 ## Usage
 
 ```go
-import "go.mongodb.org/realm/realm"
+import "go.mongodb.org/atlas-appservices/appservices"
 ```
 
-Construct a new Realm client, then use the various services on the client to
+Construct a new App Services client, then use the various services on the client to
 access different parts of the Atlas API. For example:
 
 ```go
-client := realm.NewClient(nil)
+client := appservices.NewClient(nil)
 ```
 
 The services of a client divide the API into logical chunks and correspond to
 the structure of the Atlas API documentation at
-https://docs.mongodb.com/realm/admin/api/v3/.
+https://www.mongodb.com/docs/atlas/app-services/admin/api/v3/.
 
 **NOTE:** Using the [context](https://godoc.org/context) package, one can easily
 pass cancellation signals and deadlines to various services of the client for
@@ -33,7 +31,7 @@ Each version of the client is tagged, and the version is updated accordingly.
 
 To see the list of past versions, run `git tag`.
 
-To release a new version, first ensure that [Version](./realm/realm.go) is updated 
+To release a new version, first ensure that [Version](./appservices/appservices.go) is updated 
 (i.e., before running `git push origin vx.y.z`, verify that `Version=x.y.z` should match the tag being pushed to GitHub)
 
 ## Roadmap
@@ -48,4 +46,4 @@ See our [CONTRIBUTING.md](CONTRIBUTING.md) Guide.
 
 ## License
 
-`go-client-mongodb-realm` is released under the Apache 2.0 license. See [LICENSE](LICENSE)
+`go-client-mongodb-atlas-app-services` is released under the Apache 2.0 license. See [LICENSE](LICENSE)
